naive support for dry-run.
The code in this changeset is a minimal PoC that proves that the ideas
are possible, it should be treated as such and should not be used in the
real world.

This leverages the work here:

https://github.com/kubernetes-sigs/controller-runtime/issues/831

to add dry-run support to kustomize-controller. The use case for this is
purely in CI/dev.

dev story
- developer makes changes to YAML manifests managed by flux.
- developer runs `dry-run` against a specific kustomization resource and
  sees what the effect of their changes are locally.

CI story
- A change request is made that contains updates to YAML manifests
  managed by flux.
- CI merges the change into the main branch.
- CI runs `dry-run` against a preconfigure kustomization resource and
  posts the effect of the changes as a comment on the pull request. It
  is important to note that the act of merging changes and posting a
  comment are completely outside the scope of this change and the
  kustomize-controller, this is just an example.

The ideal final state for this will be in the flux cli, with helm and
non-git sources support built in, as well as support for deletion
dry-run.

Some additonal changes that should happen to the controller:

- make `finalize` a public method so it can be used to simulate
  deletions, so `flux delete --dry-run foo` becomes possible.
- make `ExternalEventRecorder` an interface, that way the events are
  logged as the dry run output rather than relying on the controller's
  log output.

## FAQ

- Why not implement this completely outside the controller?
  The `Reconcile` method has changed a bunch of times and may be
  subject to change, hence it is important that the dry-run logic
  completely simulates the controller behavior.
- Why not use `kustomize build . | kubectl apply --dry-run=server -f -`:
  This will not account for deltas removed from the changeset as
  compared to the base, so deletions will not reflect.

+++ controllers/kustomization_controller.go
@@ -84,6 +84,8 @@ type KustomizationReconciler struct {
 	MetricsRecorder       *metrics.Recorder
 	StatusPoller          *polling.StatusPoller
 	ControllerName        string
+	SourcePath            string
+	Revision              string
 }
 
 type KustomizationReconcilerOptions struct {
@@ -173,6 +175,7 @@ func (r *KustomizationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 
 	// resolve source reference
 	source, err := r.getSource(ctx, kustomization)
+
 	if err != nil {
 		if apierrors.IsNotFound(err) {
 			msg := fmt.Sprintf("Source '%s' not found", kustomization.Spec.SourceRef.String())
@@ -191,7 +194,7 @@ func (r *KustomizationReconciler) Reconcile(ctx context.Context, req ctrl.Reques
 		}
 	}
 
-	if source.GetArtifact() == nil {
+	if source.GetArtifact() == nil && r.SourcePath != "" {
 		msg := "Source is not ready, artifact not found"
 		kustomization = kustomizev1.KustomizationNotReady(kustomization, "", kustomizev1.ArtifactFailedReason, msg)
 		if err := r.patchStatus(ctx, req, kustomization.Status); err != nil {
@@ -280,30 +283,34 @@ func (r *KustomizationReconciler) reconcile(
 		kustomization.Status.SetLastHandledReconcileRequest(v)
 	}
 
-	revision := source.GetArtifact().Revision
+	tmpDir := r.SourcePath
+	revision := r.Revision
+	if r.SourcePath == "" {
+		revision := source.GetArtifact().Revision
 
-	// create tmp dir
-	tmpDir, err := os.MkdirTemp("", kustomization.Name)
-	if err != nil {
-		err = fmt.Errorf("tmp dir error: %w", err)
-		return kustomizev1.KustomizationNotReady(
-			kustomization,
-			revision,
-			sourcev1.StorageOperationFailedReason,
-			err.Error(),
-		), err
-	}
-	defer os.RemoveAll(tmpDir)
+		// create tmp dir
+		tmpDir, err := os.MkdirTemp("", kustomization.Name)
+		if err != nil {
+			err = fmt.Errorf("tmp dir error: %w", err)
+			return kustomizev1.KustomizationNotReady(
+				kustomization,
+				revision,
+				sourcev1.StorageOperationFailedReason,
+				err.Error(),
+			), err
+		}
+		defer os.RemoveAll(tmpDir)
 
-	// download artifact and extract files
-	err = r.download(source.GetArtifact(), tmpDir)
-	if err != nil {
-		return kustomizev1.KustomizationNotReady(
-			kustomization,
-			revision,
-			kustomizev1.ArtifactFailedReason,
-			err.Error(),
-		), err
+		// download artifact and extract files
+		err = r.download(source.GetArtifact(), tmpDir)
+		if err != nil {
+			return kustomizev1.KustomizationNotReady(
+				kustomization,
+				revision,
+				kustomizev1.ArtifactFailedReason,
+				err.Error(),
+			), err
+		}
 	}
 
 	// check build path exists
@@ -579,6 +586,12 @@ func (r *KustomizationReconciler) getSource(ctx context.Context, kustomization k
 			}
 			return source, fmt.Errorf("unable to get source '%s': %w", namespacedName, err)
 		}
+		if r.SourcePath != "" {
+			a := &sourcev1.Artifact{}
+			a.Path = r.SourcePath
+			a.Revision = r.Revision
+			repository.Status.Artifact = a
+		}
 		source = &repository
 	case sourcev1.BucketKind:
 		var bucket sourcev1.Bucket
@@ -589,11 +602,18 @@ func (r *KustomizationReconciler) getSource(ctx context.Context, kustomization k
 			}
 			return source, fmt.Errorf("unable to get source '%s': %w", namespacedName, err)
 		}
+		if r.SourcePath != "" {
+			a := &sourcev1.Artifact{}
+			a.Path = r.SourcePath
+			a.Revision = r.Revision
+			bucket.Status.Artifact = a
+		}
 		source = &bucket
 	default:
 		return source, fmt.Errorf("source `%s` kind '%s' not supported",
 			kustomization.Spec.SourceRef.Name, kustomization.Spec.SourceRef.Kind)
 	}
+
 	return source, nil
 }
 
@@ -721,11 +741,13 @@ func (r *KustomizationReconciler) apply(ctx context.Context, manager *ssa.Resour
 			}
 		}
 
-		if err := manager.Wait(stageOne, ssa.WaitOptions{
-			Interval: 2 * time.Second,
-			Timeout:  kustomization.GetTimeout(),
-		}); err != nil {
-			return false, nil, err
+		if r.SourcePath == "" {
+			if err := manager.Wait(stageOne, ssa.WaitOptions{
+				Interval: 2 * time.Second,
+				Timeout:  kustomization.GetTimeout(),
+			}); err != nil {
+				return false, nil, err
+			}
 		}
 	}
